requirements.txt
Flask==2.3.3
flask-restx==1.3.2
openai>=1.0.0
httpx>=0.24.0
requests==2.31.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
flask-jwt-extended==4.6.0
PyJWT==2.10.1
bcrypt==4.1.2
Flask-Mail==0.9.1
sib-api-v3-sdk
ollama
Pillow
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
